requirements.txt
django==1.11.8
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-pure-pagination==0.3.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
celery==4.2.0
django-filters==0.2.1
django-filter==1.1.0
channels==1.1.8
future==0.16.0
django-reversion==3.0.0
django-celery-beat==1.1.0
django-celery-results==1.0.1
django-formtools==2.1
httplib2==0.9.2
django-import-export==1.0.1
asgi-redis==1.4.3
paramiko==2.4.2
ansible==2.4.1.0
coreapi==2.3.3
python-magic==0.4.3
Pillow==5.2.0
pycrypto==2.6.1
cryptography==2.5
Markdown==3.0.1
flower==0.9.3
pymysql==0.9.3
django-auth-ldap==1.7.0
django-redis==4.10.0
kubernetes==9.0.0
